:zap:UPDATE: accounts 앱과 연결, movieinfo 측에서 유저 선호도 정보 저장, 한줄평의 유저 정보 저장

+++ movieinfo/views.py
@@ -5,26 +5,24 @@ import datetime
 from django.db.models import Q
 from rest_framework import viewsets
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
 from . import utils
-from .models import (
-    MovieInfo,
-    OneLineCritic,
-    GPTAnalysis,
-    TestLikeMovie,
-    TestWatchedMovie,
-    TestWatchlistMovie,
-)
+from .models import MovieInfo, OneLineCritic, GPTAnalysis
+from .permissions import OneLineCriticIsAuthorOrReadOnly
+from accounts.models import User
+from accounts.models import LikeMovie, WatchedMovie, WatchlistMovie
+
 from .serializers import (
     MovieInfoSerializers,
     OneLineCriticSerializers,
     OneLineCriticCreateUpdateSerializers,
     GPTAnalysisSerializers,
-    TestLikeMovieSerializers,
-    TestWahtchlistMovieSerializers,
-    TestWatchedMovieSerializers,
+    LikeMovieSerializers,
+    WahtchlistMovieSerializers,
+    WatchedMovieSerializers,
 )
 from .detail_summary_with_GPT import send_data_to_GPT as GPT
 
@@ -46,9 +44,17 @@ class SerachMovieAPIView(APIView):
         return Response(serializer.data)
 
 
+class MovieInfoViewSet(viewsets.ModelViewSet):
+    queryset = MovieInfo.objects.all()
+    serializer_class = MovieInfoSerializers
+    http_method_names = ["get"]  # allowed methods
+
+
 class OneLineCriticViewSet(viewsets.ModelViewSet):
     queryset = OneLineCritic.objects.all()
     serializer_class = OneLineCriticSerializers
+    permission_classes = [OneLineCriticIsAuthorOrReadOnly]
+    http_method_names = ["get", "post", "delete", "patch"]
 
     def list(self, request, *args, **kwargs):
         pk = self.kwargs["movie_id"]
@@ -57,7 +63,10 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def create(self, request, *args, **kwargs):
-        serializer = OneLineCriticCreateUpdateSerializers(data=request.data)
+        data = request.data
+        data["author"] = User.objects.get(username=request.user).pk
+        data["movie"] = self.kwargs["movie_id"]
+        serializer = self.get_serializer(data=data)
         serializer.is_valid(raise_exception=True)
         self.perform_create(serializer)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -74,10 +83,13 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         req = self.request.data
         content = req["content"]
         starpoint = req["starpoint"]
+        user = User.objects.get(id=req["author"])
+        movie = MovieInfo.objects.get(id=self.kwargs["movie_id"])
         serializer.save(
             content=content,
             starpoint=starpoint,
-            movie=MovieInfo.objects.get(id=self.kwargs["movie_id"]),
+            movie=movie,
+            author=user,
         )
 
     def retrieve(self, request, *args, **kwargs):
@@ -105,10 +117,12 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         pk = self.kwargs["pk"]
         movie_id = self.kwargs["movie_id"]
         try:
-            instance = OneLineCritic.objects.get(Q(id=pk) & Q(movie__id=movie_id))
+            instance = OneLineCritic.objects.get(Q(id=pk) & Q(movie=movie_id))
             data = request.data
+            data["user"] = request.user
+            data["movie_id"] = movie_id
             serializer = OneLineCriticCreateUpdateSerializers(
-                instance=instance, data=data, partial=True
+                instance=instance, data=data
             )
             serializer.is_valid(raise_exception=True)
             self.perform_update(serializer)
@@ -124,16 +138,22 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         serializer.save(
             content=content,
             starpoint=starpoint,
-            movie=MovieInfo.objects.get(id=self.kwargs["movie_id"]),
         )
 
     def destroy(self, request, *args, **kwargs):
-        return super().destroy(request, *args, **kwargs)
+        pk = self.kwargs["pk"]
+        movie_id = self.kwargs["movie_id"]
 
-
-class MovieInfoViewSet(viewsets.ModelViewSet):
-    queryset = MovieInfo.objects.all()
-    serializer_class = MovieInfoSerializers
+        try:
+            instance = OneLineCritic.objects.get(Q(id=pk) & Q(movie=movie_id))
+            if request.user != instance.author:
+                raise ValueError
+        except:
+            errorMessage = {"message": "잘못된 응답입니다."}
+            return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
+        instance.delete()
+        message = {"message": "삭제되었습니다."}
+        return Response(message, status=status.HTTP_204_NO_CONTENT)
 
 
 class UserLWWViewSet(viewsets.ModelViewSet):
@@ -142,88 +162,107 @@ class UserLWWViewSet(viewsets.ModelViewSet):
 
     """
 
-    queryset = TestLikeMovie.objects.all()
-    serializer_class = TestLikeMovieSerializers
-    # authentication_classes = ()
+    queryset = LikeMovie.objects.all()
+    serializer_class = LikeMovieSerializers
+    permission_classes = [IsAuthenticated]
 
     def list(self, request, *args, **kwargs):
         pk = self.kwargs["movie_id"]
         mode = kwargs["mode"]
 
         if mode == "like":
-            queryset = TestLikeMovie.objects.filter(movie__id=pk)
-            serializer = TestLikeMovieSerializers(queryset, many=True)
+            queryset = LikeMovie.objects.filter(movie__id=pk)
+            serializer = LikeMovieSerializers(queryset, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         elif mode == "watchlist":
-            queryset = TestWatchlistMovie.objects.filter(movie__id=pk)
-            serializer = TestWahtchlistMovieSerializers(queryset, many=True)
+            queryset = WatchlistMovie.objects.filter(movie__id=pk)
+            serializer = WahtchlistMovieSerializers(queryset, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         elif mode == "watched":
-            queryset = TestWatchedMovie.objects.filter(movie__id=pk)
-            serializer = TestWatchedMovieSerializers(queryset, many=True)
+            queryset = WatchedMovie.objects.filter(movie__id=pk)
+            serializer = WatchedMovieSerializers(queryset, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         else:
             errorMessage = {"message": "잘못된 응답입니다."}
             return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
 
+    def retrieve(self, request, *args, **kwargs):
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
     def create(self, request, *args, **kwargs):
         movie_id = kwargs["movie_id"]
         mode = kwargs["mode"]
+        user = request.user
+        data = {"movie_id": movie_id, "user": user}
         if mode == "like":
-            instance = TestLikeMovieSerializers(data={"movie_id": movie_id})
+            instance = LikeMovieSerializers(data=data)
             instance.is_valid(raise_exception=True)
-            instance.save(movie=MovieInfo.objects.get(id=movie_id))
+            instance.save(
+                movie=MovieInfo.objects.get(id=movie_id), user=self.request.user
+            )
             return Response(instance.data, status=status.HTTP_201_CREATED)
         elif mode == "watchlist":
-            instance = TestWahtchlistMovieSerializers(data={"movie_id": movie_id})
+            instance = WahtchlistMovieSerializers(data=data)
             instance.is_valid(raise_exception=True)
-            instance.save(movie=MovieInfo.objects.get(id=movie_id))
+            instance.save(
+                movie=MovieInfo.objects.get(id=movie_id), user=self.request.user
+            )
             return Response(instance.data, status=status.HTTP_201_CREATED)
         elif mode == "watched":
-            instance = TestWatchedMovieSerializers(data={"movie_id": movie_id})
+            instance = WatchedMovieSerializers(data=data)
             instance.is_valid(raise_exception=True)
-            instance.save(movie=MovieInfo.objects.get(id=movie_id))
+            instance.save(
+                movie=MovieInfo.objects.get(id=movie_id), user=self.request.user
+            )
             return Response(instance.data, status=status.HTTP_201_CREATED)
         else:
             errorMessage = {"message": "잘못된 응답입니다."}
             return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
 
     def destroy(self, request, *args, **kwargs):
+        pk = self.kwargs["pk"]
         movie_id = kwargs["movie_id"]
         mode = kwargs["mode"]
+        user = request.user
         if mode == "like":
             try:
-                instance = TestLikeMovie.objects.get(
-                    movie=MovieInfo.objects.get(id=movie_id)
+                instance = LikeMovie.objects.get(
+                    Q(id=pk) & Q(movie=MovieInfo.objects.get(id=movie_id))
                 )
-                instance.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
+                if user == instance.user:
+                    instance.delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
+                else:
+                    raise ValueError
             except:
                 errorMessage = {"message": "잘못된 응답입니다."}
                 return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
         elif mode == "watchlist":
             try:
-                instance = TestWatchlistMovie.objects.get(
-                    movie=MovieInfo.objects.get(id=movie_id)
+                instance = WatchlistMovie.objects.get(
+                    Q(id=pk) & Q(movie=MovieInfo.objects.get(id=movie_id))
                 )
-                instance.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
+                if user == instance.user:
+                    instance.delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
+                else:
+                    raise ValueError
             except:
                 errorMessage = {"message": "잘못된 응답입니다."}
                 return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
         elif mode == "watched":
             try:
-                instance = TestWatchedMovie.objects.get(
-                    movie=MovieInfo.objects.get(id=movie_id)
+                instance = WatchedMovie.objects.get(
+                    Q(id=pk) & Q(movie=MovieInfo.objects.get(id=movie_id))
                 )
-                instance.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
+                if user == instance.user:
+                    instance.delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
+                else:
+                    raise ValueError
             except:
                 errorMessage = {"message": "잘못된 응답입니다."}
                 return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
-        else:
-            errorMessage = {"message": "잘못된 응답입니다."}
-            return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
 
 
 class GPTAnalysisViewSet(viewsets.ModelViewSet):
